+++ GiftCraft-BackEnd/src/main/java/com/example/GiftCraft_BackEnd/Repo/Cart/CartRepo.java
@@ -7,5 +7,5 @@ import java.util.Optional;
 
 public interface CartRepo extends JpaRepository<Cart, Integer> {
 
-    Optional<Cart> findByUserId(int userId);
+    Optional<Cart> findByUserId(Long userId);
 }
